fix: coderabbit

+++ src/main/java/com/example/Dingle/environment/repository/ParticulateMatterRepository.java
@@ -5,6 +5,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface ParticulateMatterRepository extends JpaRepository<ParticulateMatter, Integer> {
+public interface ParticulateMatterRepository extends JpaRepository<ParticulateMatter, Long> {
     ParticulateMatter findByDistrict_Id(Long districtId);
 }
